update tests (#446)
This pull request refactors the end-to-end (E2E) test suite to improve
test organization, coverage, and maintainability. It restructures the
test matrix in the GitHub Actions workflow to simplify configuration,
raises the code coverage threshold, and introduces new and updated tests
for IBC functionality and DKIM key validation. Several test files have
been renamed and moved to better reflect their module boundaries.

**Test Suite Restructuring and Workflow Improvements**
- Consolidated the E2E test matrix in `.github/workflows/e2e-tests.yaml`
by grouping tests by module and type, replacing the large flat list with
a more maintainable structure. This change also updates the test runner
to pass both test type and directory, improving flexibility for future
additions.
- Updated Docker image tagging and usage in the workflow to support
upgrade tests, ensuring the correct local image is used during test
execution.

**Coverage and Quality**
- Increased the code coverage threshold from 70% to 80% in `.coveragerc`
to enforce higher test coverage standards.

**IBC Module Test Enhancements**
- Moved XionMinFeeIBC test to `ibc_min_fee_test.go` 
- Improved and renamed the IBC timeout test (`ibc_timeout_test.go`),
including better handling of relayer state, more robust timeout
scenarios, and clearer refund logic. The test now covers both height and
timestamp-based timeouts and includes detailed comments and checks for
correct token refunding.
[[1]](diffhunk://#diff-becbffaf47b368ebc31b0e43a0664517a383b6b9f12ad004ae417dcf7510474cL1-L8)
[[2]](diffhunk://#diff-becbffaf47b368ebc31b0e43a0664517a383b6b9f12ad004ae417dcf7510474cL30-R29)
[[3]](diffhunk://#diff-becbffaf47b368ebc31b0e43a0664517a383b6b9f12ad004ae417dcf7510474cR107-R117)
[[4]](diffhunk://#diff-becbffaf47b368ebc31b0e43a0664517a383b6b9f12ad004ae417dcf7510474cL136-R175)
[[5]](diffhunk://#diff-becbffaf47b368ebc31b0e43a0664517a383b6b9f12ad004ae417dcf7510474cL163-R202)

**DKIM Module Test Improvements**
- Refactored the DKIM test to correctly calculate the decoded size of
base64 public keys for size validation, ensuring the test matches the
actual parameter logic and preventing false positives/negatives.
[[1]](diffhunk://#diff-075f1fabaf297ea005b8651973536c29df694ecc8d4ab4e04be8338a2baeeaffL1-R1)
[[2]](diffhunk://#diff-075f1fabaf297ea005b8651973536c29df694ecc8d4ab4e04be8338a2baeeaffL87-R96)

**Test File Organization**
- Renamed and moved several test files to reflect their module
boundaries, improving clarity and future maintainability (e.g., moving
`webauthn_test.go` and `genesis_export_import_test.go` to their
respective modules).
[[1]](diffhunk://#diff-c4fe28d64d77cbd5078203772530d62aad17a712c73400df09382ed593d32084L1-R1)
[[2]](diffhunk://#diff-c3359fe8661dbda6f8e26caa721c03b2594b4ac7121e426f186b6f9dc76442f9L1-R1)

This refactor improves the reliability, maintainability, and coverage of
the E2E test suite, while also making the CI workflow easier to extend
and understand.
