+++ e2e_tests/abstract-account/webauthn_test.go
@@ -1,4 +1,4 @@
-package e2e_xion
+package e2e_aa
 
 import (
 	"crypto/sha256"
