+++ e2e_tests/app/genesis_export_import_test.go
@@ -1,4 +1,4 @@
-package e2e_xion
+package e2e_app
 
 import (
 	"context"
