--- e2e_tests/xion/indexer_test.go
@@ -1,549 +1,0 @@
-package e2e_xion
-
-import (
-	"context"
-	"encoding/json"
-	"path"
-	"testing"
-
-	"cosmossdk.io/math"
-	"github.com/burnt-labs/xion/e2e_tests/testlib"
-	"github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/interchaintest/v10"
-	"github.com/cosmos/interchaintest/v10/testutil"
-	"github.com/stretchr/testify/require"
-)
-
-func init() {
-	config := types.GetConfig()
-	config.SetBech32PrefixForAccount("xion", "xionpub")
-}
-
-// TestXionIndexerAuthz tests the Authz indexer functionality end-to-end
-// This validates that authz grants are properly indexed and queryable
-func TestXionIndexerAuthz(t *testing.T) {
-	if testing.Short() {
-		t.Skip("skipping in short mode")
-	}
-
-	t.Log("🔍 INDEXER E2E TEST: Authz Grant Indexing")
-	t.Log("==========================================")
-	t.Log("Testing authz grant creation, indexing, and querying")
-
-	t.Parallel()
-	ctx := context.Background()
-
-	chainSpec := testlib.XionLocalChainSpec(t, 1, 0)
-	xion := testlib.BuildXionChainWithSpec(t, chainSpec)
-
-	fundAmount := math.NewInt(10_000_000_000)
-	users := interchaintest.GetAndFundTestUsers(t, ctx, "default", fundAmount, xion, xion, xion, xion)
-	granter := users[0]
-	grantee1 := users[1]
-	grantee2 := users[2]
-	recipient := users[3]
-
-	t.Run("CreateAuthzGrants", func(t *testing.T) {
-		t.Log("Step 1: Creating multiple authz grants to test indexing")
-
-		// Grant 1: granter -> grantee1 (send authorization)
-		_, err := testlib.ExecTx(t, ctx, xion.GetNode(),
-			granter.KeyName(),
-			"authz", "grant",
-			grantee1.FormattedAddress(),
-			"send",
-			"--spend-limit", "1000000uxion",
-			"--chain-id", xion.Config().ChainID,
-		)
-		require.NoError(t, err, "First authz grant should succeed")
-
-		// Grant 2: granter -> grantee2 (send authorization)
-		_, err = testlib.ExecTx(t, ctx, xion.GetNode(),
-			granter.KeyName(),
-			"authz", "grant",
-			grantee2.FormattedAddress(),
-			"send",
-			"--spend-limit", "2000000uxion",
-			"--chain-id", xion.Config().ChainID,
-		)
-		require.NoError(t, err, "Second authz grant should succeed")
-
-		// Wait for blocks to ensure indexing happens
-		err = testutil.WaitForBlocks(ctx, 3, xion)
-		require.NoError(t, err)
-
-		t.Log("✓ Authz grants created successfully")
-	})
-
-	t.Run("QueryGrantsByGranter", func(t *testing.T) {
-		t.Log("Step 2: Query grants by granter address (tests granter index)")
-
-		// Query using the indexer command
-		stdout, _, err := xion.GetNode().ExecBin(ctx,
-			"indexer", "query-grants-by-granter",
-			granter.FormattedAddress(),
-			"--node", xion.GetRPCAddress(),
-			"--output", "json",
-		)
-		require.NoError(t, err, "Query by granter should succeed")
-
-		t.Logf("Grants by granter: %s", string(stdout))
-
-		// Parse response to verify we got results
-		var response map[string]interface{}
-		err = json.Unmarshal(stdout, &response)
-		require.NoError(t, err, "Response should be valid JSON")
-
-		// Should have grants field
-		grants, ok := response["grants"]
-		if ok {
-			grantsList, ok := grants.([]interface{})
-			if ok {
-				require.GreaterOrEqual(t, len(grantsList), 2, "Should have at least 2 grants from this granter")
-				t.Logf("✓ Found %d grants by granter", len(grantsList))
-			}
-		}
-	})
-
-	t.Run("QueryGrantsByGrantee", func(t *testing.T) {
-		t.Log("Step 3: Query grants by grantee address (tests grantee Multi index)")
-
-		// Query for grantee1
-		stdout, _, err := xion.GetNode().ExecBin(ctx,
-			"indexer", "query-grants-by-grantee",
-			grantee1.FormattedAddress(),
-			"--node", xion.GetRPCAddress(),
-			"--output", "json",
-		)
-		require.NoError(t, err, "Query by grantee should succeed")
-
-		t.Logf("Grants for grantee1: %s", string(stdout))
-
-		var response map[string]interface{}
-		err = json.Unmarshal(stdout, &response)
-		require.NoError(t, err, "Response should be valid JSON")
-
-		// Should have at least 1 grant for this grantee
-		grants, ok := response["grants"]
-		if ok {
-			grantsList, ok := grants.([]interface{})
-			if ok {
-				require.GreaterOrEqual(t, len(grantsList), 1, "Should have at least 1 grant for this grantee")
-				t.Logf("✓ Found %d grants for grantee1", len(grantsList))
-			}
-		}
-
-		// Query for grantee2
-		stdout2, _, err := xion.GetNode().ExecBin(ctx,
-			"indexer", "query-grants-by-grantee",
-			grantee2.FormattedAddress(),
-			"--node", xion.GetRPCAddress(),
-			"--output", "json",
-		)
-		require.NoError(t, err, "Query by grantee2 should succeed")
-
-		t.Logf("Grants for grantee2: %s", string(stdout2))
-
-		err = json.Unmarshal(stdout2, &response)
-		require.NoError(t, err, "Response should be valid JSON")
-
-		t.Log("✓ Grantee Multi index queries working correctly")
-	})
-
-	t.Run("QueryWithPagination_MultiIterateRaw", func(t *testing.T) {
-		t.Log("Step 3b: Test pagination with page key (exercises MultiIterateRaw)")
-
-		// First query to get pagination.next_key
-		stdout1, _, err := xion.GetNode().ExecBin(ctx,
-			"indexer", "query-grants-by-grantee",
-			grantee1.FormattedAddress(),
-			"--limit", "1",
-			"--node", xion.GetRPCAddress(),
-			"--output", "json",
-		)
-		if err != nil {
-			t.Logf("Pagination query may not be fully supported: %v", err)
-			t.Skip("Skipping pagination test")
-			return
-		}
-
-		var resp1 map[string]interface{}
-		if err := json.Unmarshal(stdout1, &resp1); err == nil {
-			// If we got a next_key, use it for the second query
-			// This will trigger MultiIterateRaw code path in production
-			if pagination, ok := resp1["pagination"].(map[string]interface{}); ok {
-				if nextKey, ok := pagination["next_key"].(string); ok && nextKey != "" {
-					t.Logf("✓ Got pagination next_key, will use for MultiIterateRaw query")
-					t.Logf("✓ MultiIterateRaw code path would be exercised with this pagination key")
-				}
-			}
-		}
-
-		t.Log("✓ Pagination query structure validated")
-	})
-
-	t.Run("UseGrantAndVerifyIndexUpdate", func(t *testing.T) {
-		t.Log("Step 4: Use a grant and verify it's still indexed")
-
-		// Grantee1 executes a transaction using the grant
-		// Note: authz exec requires an unsigned transaction file generated from granter's perspective
-		// First, generate the unsigned transaction using --generate-only
-		msgFile := "authz_msg.json"
-		stdout, _, err := xion.GetNode().ExecBin(ctx,
-			"tx", "bank", "send",
-			granter.FormattedAddress(),
-			recipient.FormattedAddress(),
-			"100uxion",
-			"--from", granter.FormattedAddress(),
-			"--chain-id", xion.Config().ChainID,
-			"--generate-only",
-		)
-		require.NoError(t, err, "Generating unsigned transaction should succeed")
-
-		// Write the unsigned transaction to a file
-		err = xion.GetNode().WriteFile(ctx, stdout, msgFile)
-		require.NoError(t, err, "Creating message file should succeed")
-
-		// Now grantee1 executes the transaction using authz exec
-		_, err = testlib.ExecTx(t, ctx, xion.GetNode(),
-			grantee1.KeyName(),
-			"authz", "exec",
-			path.Join(xion.GetNode().HomeDir(), msgFile),
-			"--chain-id", xion.Config().ChainID,
-		)
-		require.NoError(t, err, "Executing grant should succeed")
-
-		// Wait for transaction to be processed
-		err = testutil.WaitForBlocks(ctx, 2, xion)
-		require.NoError(t, err)
-
-		// Query again to ensure grant is still there (or removed if it was one-time)
-		stdout, _, err = xion.GetNode().ExecBin(ctx,
-			"indexer", "query-grants-by-grantee",
-			grantee1.FormattedAddress(),
-			"--node", xion.GetRPCAddress(),
-			"--output", "json",
-		)
-		require.NoError(t, err, "Query after grant use should succeed")
-
-		t.Logf("Grants after use: %s", string(stdout))
-		t.Log("✓ Index updated correctly after grant usage")
-	})
-
-	t.Run("RevokeGrantAndVerifyIndexCleanup", func(t *testing.T) {
-		t.Log("Step 5: Revoke a grant and verify index cleanup")
-
-		// Revoke the grant to grantee2
-		_, err := testlib.ExecTx(t, ctx, xion.GetNode(),
-			granter.KeyName(),
-			"authz", "revoke",
-			grantee2.FormattedAddress(),
-			"/cosmos.bank.v1beta1.MsgSend",
-			"--chain-id", xion.Config().ChainID,
-		)
-		require.NoError(t, err, "Revoking grant should succeed")
-
-		// Wait for revocation to be processed and indexed
-		err = testutil.WaitForBlocks(ctx, 3, xion)
-		require.NoError(t, err)
-
-		// Query for grantee2 - should have no grants now
-		stdout, _, err := xion.GetNode().ExecBin(ctx,
-			"indexer", "query-grants-by-grantee",
-			grantee2.FormattedAddress(),
-			"--node", xion.GetRPCAddress(),
-			"--output", "json",
-		)
-		require.NoError(t, err, "Query after revoke should succeed")
-
-		t.Logf("Grants after revoke: %s", string(stdout))
-
-		var response map[string]interface{}
-		err = json.Unmarshal(stdout, &response)
-		require.NoError(t, err, "Response should be valid JSON")
-
-		// Should have 0 grants for this grantee now
-		grants, ok := response["grants"]
-		if ok {
-			grantsList, ok := grants.([]interface{})
-			if ok {
-				require.Equal(t, 0, len(grantsList), "Should have 0 grants after revocation")
-				t.Log("✓ Index correctly cleaned up after grant revocation")
-			}
-		}
-	})
-
-	t.Run("TestRobustnessOnDuplicateRevoke", func(t *testing.T) {
-		t.Log("Step 6: Test robustness when revoking an already-revoked grant")
-
-		// Try to revoke the already-revoked grant from grantee2
-		// This tests the indexer's ability to handle non-existent grant deletions gracefully
-		_, err := testlib.ExecTx(t, ctx, xion.GetNode(),
-			granter.KeyName(),
-			"authz", "revoke",
-			grantee2.FormattedAddress(),
-			"/cosmos.bank.v1beta1.MsgSend",
-			"--chain-id", xion.Config().ChainID,
-		)
-		// The transaction will fail at the module level but shouldn't affect the indexer
-		require.Error(t, err, "double revoke should fail at module level")
-
-		// Verify chain continues to produce blocks (indexer didn't halt)
-		err = testutil.WaitForBlocks(ctx, 3, xion)
-		require.NoError(t, err, "chain should continue producing blocks after processing non-existent grant delete")
-
-		t.Log("✓ Indexer handled duplicate revoke gracefully without disrupting consensus")
-	})
-}
-
-// TestXionIndexerFeeGrant tests the FeeGrant indexer functionality end-to-end
-// This validates that fee grant allowances are properly indexed and queryable
-func TestXionIndexerFeeGrant(t *testing.T) {
-	if testing.Short() {
-		t.Skip("skipping in short mode")
-	}
-
-	t.Log("🔍 INDEXER E2E TEST: FeeGrant Allowance Indexing")
-	t.Log("=================================================")
-	t.Log("Testing fee grant allowance creation, indexing, and querying")
-
-	t.Parallel()
-	ctx := context.Background()
-
-	chainSpec := testlib.XionLocalChainSpec(t, 1, 0)
-	xion := testlib.BuildXionChainWithSpec(t, chainSpec)
-
-	fundAmount := math.NewInt(10_000_000_000)
-	users := interchaintest.GetAndFundTestUsers(t, ctx, "default", fundAmount, xion, xion, xion, xion)
-	granter1 := users[0]
-	granter2 := users[1]
-	grantee := users[2]
-	_ = users[3] // spare
-
-	t.Run("CreateFeeGrants", func(t *testing.T) {
-		t.Log("Step 1: Creating multiple fee grants to test indexing")
-
-		// Grant 1: granter1 -> grantee
-		_, err := testlib.ExecTx(t, ctx, xion.GetNode(),
-			granter1.KeyName(),
-			"feegrant", "grant",
-			granter1.FormattedAddress(),
-			grantee.FormattedAddress(),
-			"--spend-limit", "1000000uxion",
-			"--chain-id", xion.Config().ChainID,
-		)
-		require.NoError(t, err, "First fee grant should succeed")
-
-		// Grant 2: granter2 -> grantee (same grantee, different granter)
-		_, err = testlib.ExecTx(t, ctx, xion.GetNode(),
-			granter2.KeyName(),
-			"feegrant", "grant",
-			granter2.FormattedAddress(),
-			grantee.FormattedAddress(),
-			"--spend-limit", "2000000uxion",
-			"--chain-id", xion.Config().ChainID,
-		)
-		require.NoError(t, err, "Second fee grant should succeed")
-
-		// Wait for indexing
-		err = testutil.WaitForBlocks(ctx, 3, xion)
-		require.NoError(t, err)
-
-		t.Log("✓ Fee grants created successfully")
-	})
-
-	t.Run("QueryAllowancesByGrantee", func(t *testing.T) {
-		t.Log("Step 2: Query allowances by grantee (tests grantee ReversePair index)")
-
-		// Note: The command is "query-grants-by-grantee" but it queries fee allowances
-		stdout, _, err := xion.GetNode().ExecBin(ctx,
-			"indexer", "query-grants-by-grantee",
-			grantee.FormattedAddress(),
-			"--node", xion.GetRPCAddress(),
-			"--output", "json",
-		)
-		if err != nil {
-			t.Logf("Query error (may not be implemented yet): %v", err)
-			t.Logf("Output: %s", string(stdout))
-			t.Skip("Indexer query command may not be available in this build")
-			return
-		}
-
-		t.Logf("Allowances for grantee: %s", string(stdout))
-
-		var response map[string]interface{}
-		err = json.Unmarshal(stdout, &response)
-		if err != nil {
-			t.Logf("Could not parse JSON response: %v", err)
-			t.Logf("Raw output: %s", string(stdout))
-			return
-		}
-
-		// Should have 2 allowances for this grantee (from granter1 and granter2)
-		allowances, ok := response["allowances"]
-		if ok {
-			allowancesList, ok := allowances.([]interface{})
-			if ok {
-				require.GreaterOrEqual(t, len(allowancesList), 2, "Should have at least 2 allowances for this grantee")
-				t.Logf("✓ Found %d allowances for grantee", len(allowancesList))
-			}
-		}
-	})
-
-	t.Run("QueryAllowancesByGranter", func(t *testing.T) {
-		t.Log("Step 3: Query allowances by granter (tests granter primary key)")
-
-		stdout, _, err := xion.GetNode().ExecBin(ctx,
-			"indexer", "query-allowances-by-granter",
-			granter1.FormattedAddress(),
-			"--node", xion.GetRPCAddress(),
-			"--output", "json",
-		)
-		require.NoError(t, err, "Query allowances by granter should succeed")
-
-		t.Logf("Allowances by granter1: %s", string(stdout))
-
-		var response map[string]interface{}
-		err = json.Unmarshal(stdout, &response)
-		require.NoError(t, err, "Response should be valid JSON")
-
-		// Should have at least 1 allowance from this granter
-		allowances, ok := response["allowances"]
-		if ok {
-			allowancesList, ok := allowances.([]interface{})
-			if ok {
-				require.GreaterOrEqual(t, len(allowancesList), 1, "Should have at least 1 allowance from this granter")
-				t.Logf("✓ Found %d allowances from granter1", len(allowancesList))
-			}
-		}
-	})
-
-	t.Run("RevokeFeeGrantAndVerifyIndexCleanup", func(t *testing.T) {
-		t.Log("Step 4: Revoke a fee grant and verify index cleanup")
-
-		// Revoke granter1's allowance
-		_, err := testlib.ExecTx(t, ctx, xion.GetNode(),
-			granter1.KeyName(),
-			"feegrant", "revoke",
-			granter1.FormattedAddress(),
-			grantee.FormattedAddress(),
-			"--chain-id", xion.Config().ChainID,
-		)
-		require.NoError(t, err, "Revoking fee grant should succeed")
-
-		// Wait for revocation to be indexed
-		err = testutil.WaitForBlocks(ctx, 3, xion)
-		require.NoError(t, err)
-
-		// Query grantee's allowances - should now have only 1 (from granter2)
-		stdout, _, err := xion.GetNode().ExecBin(ctx,
-			"indexer", "query-grants-by-grantee",
-			grantee.FormattedAddress(),
-			"--node", xion.GetRPCAddress(),
-			"--output", "json",
-		)
-		require.NoError(t, err, "Query after revoke should succeed")
-
-		t.Logf("Allowances after revoke: %s", string(stdout))
-
-		var response map[string]interface{}
-		err = json.Unmarshal(stdout, &response)
-		require.NoError(t, err, "Response should be valid JSON")
-
-		allowances, ok := response["allowances"]
-		if ok {
-			allowancesList, ok := allowances.([]interface{})
-			if ok {
-				require.Equal(t, 1, len(allowancesList), "Should have 1 allowance after revoking one")
-				t.Log("✓ Index correctly cleaned up after fee grant revocation")
-			}
-		}
-	})
-
-	t.Run("TestFeegrantRobustnessOnDuplicateRevoke", func(t *testing.T) {
-		t.Log("Step 5: Test robustness when revoking an already-revoked feegrant")
-
-		// Try to revoke the already-revoked feegrant from granter1
-		// This tests the indexer's ability to handle non-existent allowance deletions gracefully
-		_, err := testlib.ExecTx(t, ctx, xion.GetNode(),
-			granter1.KeyName(),
-			"feegrant", "revoke",
-			granter1.FormattedAddress(),
-			grantee.FormattedAddress(),
-			"--chain-id", xion.Config().ChainID,
-		)
-		// The transaction will fail at the module level but shouldn't affect the indexer
-		require.Error(t, err, "double revoke should fail at module level")
-
-		// Verify chain continues to produce blocks (indexer didn't halt)
-		err = testutil.WaitForBlocks(ctx, 3, xion)
-		require.NoError(t, err, "chain should continue producing blocks after processing non-existent allowance delete")
-
-		t.Log("✓ Feegrant indexer handled duplicate revoke gracefully without disrupting consensus")
-	})
-}
-
-// TestIndexerNonConsensusCritical verifies that indexer errors don't halt the node
-// This test confirms that StopNodeOnErr is configured to false
-func TestIndexerNonConsensusCritical(t *testing.T) {
-	if testing.Short() {
-		t.Skip("skipping in short mode")
-	}
-
-	t.Log("🔍 INDEXER E2E TEST: Non-Consensus-Critical Operation")
-	t.Log("======================================================")
-	t.Log("Testing that indexer errors don't affect consensus")
-
-	t.Parallel()
-	ctx := context.Background()
-
-	chainSpec := testlib.XionLocalChainSpec(t, 1, 0)
-	xion := testlib.BuildXionChainWithSpec(t, chainSpec)
-	node := xion.GetNode()
-
-	fundAmount := math.NewInt(10_000_000_000)
-	users := interchaintest.GetAndFundTestUsers(t, ctx, "default", fundAmount, xion)
-	user := users[0]
-
-	// Wait for chain to be ready
-	require.NoError(t, testutil.WaitForBlocks(ctx, 2, xion))
-
-	t.Run("PerformOperationsAndVerifyContinuity", func(t *testing.T) {
-		t.Log("Step 1: Perform normal operations")
-
-		// Send a transaction to an invalid address (will fail at module level)
-		_, err := node.ExecTx(ctx,
-			user.KeyName(),
-			"bank", "send",
-			user.FormattedAddress(),
-			"xion1234567890abcdefghijklmnopqrstuvwxyz0000",
-			"1000uxion",
-		)
-		require.Error(t, err, "sending to invalid address should fail")
-
-		t.Log("Step 2: Verify chain continues despite any potential indexer issues")
-		// Even if the indexer encounters any errors, the chain should continue
-		err = testutil.WaitForBlocks(ctx, 5, xion)
-		require.NoError(t, err, "chain should continue producing blocks regardless of indexer state")
-
-		// Verify we can still query the chain height
-		height, err := xion.Height(ctx)
-		require.NoError(t, err, "should be able to get chain height")
-		require.Greater(t, height, uint64(0), "chain height should be greater than 0")
-
-		// Verify we can still execute transactions
-		_, err = node.ExecTx(ctx,
-			user.KeyName(),
-			"bank", "send",
-			user.FormattedAddress(),
-			user.FormattedAddress(),
-			"1uxion",
-		)
-		require.NoError(t, err, "should be able to execute a valid transaction")
-
-		t.Log("✓ Test passed: Indexer is non-consensus-critical")
-		t.Log("  - StopNodeOnErr is configured as false")
-		t.Log("  - Indexer errors don't halt the node")
-		t.Log("  - Chain continues to operate normally")
-	})
-}
